badass_api/requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
filterpy==1.4.5
Flask==1.1.1
Flask-Cors==3.0.8
marshmallow==3.3.0
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
marshmallow-sqlalchemy==0.21.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.16.4
oauthlib==3.1.0
onnx==1.6.0
onnxruntime==1.0.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
logging==0.5.1.2
catsim==0.15.4
pandas==0.25.1

